update: test_cosine_annealing_warmup_restarts
